Update all Domain entities and Fluent API configs for all entities

+++ GameIt.Domain/Discount.cs
@@ -7,12 +7,12 @@ using System.Threading.Tasks;
 using GameIt.Domain.Common;
 
 namespace GameIt.Domain
-{
+{   
     public class Discount : BaseEntity
     {
         public Guid GameId { get; set; }
         public Game Game { get; set; }
-        public int Percentage { get; set; }
+        public decimal Percentage { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
         public bool IsActive { get; set; } = true;
